requirements.txt
opensearch-py==2.8.0
python-dotenv==1.1.0
pyyaml==6.0.2
openai==1.76.2
pymongo==4.12.1
pydantic==2.11.4